+++ lib/bibliothecary/multi_parsers/cyclonedx.rb
@@ -12,17 +12,17 @@ Warning[:experimental] = true
 
 module Bibliothecary
   module MultiParsers
-    module CycloneDX
+    class CycloneDX
       include Bibliothecary::Analyser
-      include Bibliothecary::Analyser::TryCache
+      extend Bibliothecary::Analyser::TryCache
 
       NoComponents = Class.new(StandardError)
 
       class ManifestEntries
-        attr_reader :manifests
+        attr_reader :entries
 
         def initialize(parse_queue:)
-          @manifests = {}
+          @entries = Set.new
 
           # Instead of recursing, we'll work through a queue of components
           # to process, letting the different parser add components to the
@@ -31,14 +31,13 @@ module Bibliothecary
         end
 
         def add(purl, source = nil)
-          mapping = PurlUtil::PURL_TYPE_MAPPING[purl.type]
-          return unless mapping
+          # Use the mapped purl->bibliothecary platform, or else fall back to original platform itself.
+          mapping = PurlUtil::PURL_TYPE_MAPPING.fetch(purl.type, purl.type)
 
-          @manifests[mapping] ||= Set.new
-          @manifests[mapping] << Dependency.new(
+          @entries << Dependency.new(
             name: PurlUtil.full_name(purl),
             requirement: purl.version,
-            platform: mapping.to_s,
+            platform: mapping ? mapping.to_s : purl.type,
             type: "lockfile",
             source: source
           )
@@ -59,10 +58,6 @@ module Bibliothecary
 
             add(purl, source)
           end
-        end
-
-        def [](key)
-          @manifests[key]&.to_a
         end
       end
 
@@ -91,25 +86,31 @@ module Bibliothecary
         }
       end
 
-      def parse_cyclonedx_json(file_contents, options: {})
+      def self.platform_name
+        raise "CycloneDX is a multi-parser and does not have a platform name."
+      end
+
+      def self.parse_cyclonedx_json(file_contents, options: {})
         manifest = try_cache(options, options[:filename]) do
           JSON.parse(file_contents)
         end
 
         raise NoComponents unless manifest["components"]
 
-        entries = ManifestEntries.new(parse_queue: manifest["components"])
+        manifest_entries = ManifestEntries.new(
+          parse_queue: manifest["components"]
+        )
 
-        entries.parse!(options.fetch(:filename, nil)) do |component, parse_queue|
+        manifest_entries.parse!(options.fetch(:filename, nil)) do |component, parse_queue|
           parse_queue.concat(component["components"]) if component["components"]
 
           component["purl"]
         end
 
-        ParserResult.new(dependencies: entries[platform_name.to_sym] || [])
+        ParserResult.new(dependencies: manifest_entries.entries.to_a)
       end
 
-      def parse_cyclonedx_xml(file_contents, options: {})
+      def self.parse_cyclonedx_xml(file_contents, options: {})
         manifest = try_cache(options, options[:filename]) do
           Ox.parse(file_contents)
         end
@@ -121,9 +122,11 @@ module Bibliothecary
 
         raise NoComponents unless root.locate("components").first
 
-        entries = ManifestEntries.new(parse_queue: root.locate("components/*"))
+        manifest_entries = ManifestEntries.new(
+          parse_queue: root.locate("components/*")
+        )
 
-        entries.parse!(options.fetch(:filename, nil)) do |component, parse_queue|
+        manifest_entries.parse!(options.fetch(:filename, nil)) do |component, parse_queue|
           # #locate returns an empty array if nothing is found, so we can
           # always safely concatenate it to the parse queue.
           parse_queue.concat(component.locate("components/*"))
@@ -131,7 +134,7 @@ module Bibliothecary
           component.locate("purl").first&.text
         end
 
-        ParserResult.new(dependencies: entries[platform_name.to_sym] || [])
+        ParserResult.new(dependencies: manifest_entries.entries.to_a)
       end
     end
   end