+++ lib/bibliothecary/parsers/cargo.rb
@@ -18,10 +18,6 @@ module Bibliothecary
         }
       end
 
-      add_multi_parser(Bibliothecary::MultiParsers::CycloneDX)
-      add_multi_parser(Bibliothecary::MultiParsers::Spdx)
-      add_multi_parser(Bibliothecary::MultiParsers::DependenciesCSV)
-
       def self.parse_manifest(file_contents, options: {})
         manifest = Tomlrb.parse(file_contents)