+++ lib/bibliothecary/parsers/cpan.rb
@@ -21,8 +21,6 @@ module Bibliothecary
         }
       end
 
-      add_multi_parser(Bibliothecary::MultiParsers::DependenciesCSV)
-
       def self.parse_json_manifest(file_contents, options: {})
         manifest = JSON.parse file_contents
         dependencies = manifest["prereqs"].map do |_group, deps|