+++ lib/bibliothecary/parsers/elm.rb
@@ -6,7 +6,7 @@ module Bibliothecary
   module Parsers
     class Elm
       include Bibliothecary::Analyser
-      extend Bibliothecary::MultiParsers::JSONRuntime
+      extend Bibliothecary::ParserMixins::JSONRuntime
 
       def self.mapping
         {
@@ -20,8 +20,6 @@ module Bibliothecary
           },
         }
       end
-
-      add_multi_parser(Bibliothecary::MultiParsers::DependenciesCSV)
 
       def self.parse_json_lock(file_contents, options: {})
         manifest = JSON.parse file_contents