+++ lib/bibliothecary/parsers/maven.rb
@@ -133,10 +133,6 @@ module Bibliothecary
         }
       end
 
-      add_multi_parser(Bibliothecary::MultiParsers::CycloneDX)
-      add_multi_parser(Bibliothecary::MultiParsers::Spdx)
-      add_multi_parser(Bibliothecary::MultiParsers::DependenciesCSV)
-
       def self.parse_ivy_manifest(file_contents, options: {})
         manifest = Ox.parse file_contents
         dependencies = manifest.dependencies.locate("dependency").map do |dependency|