+++ lib/bibliothecary/parsers/meteor.rb
@@ -6,7 +6,7 @@ module Bibliothecary
   module Parsers
     class Meteor
       include Bibliothecary::Analyser
-      extend Bibliothecary::MultiParsers::JSONRuntime
+      extend Bibliothecary::ParserMixins::JSONRuntime
 
       def self.mapping
         {
@@ -16,8 +16,6 @@ module Bibliothecary
           },
         }
       end
-
-      add_multi_parser(Bibliothecary::MultiParsers::DependenciesCSV)
     end
   end
 end